unified maven and gradle tasks
