--- src/main/java/io/github/fvarrui/javapackager/ideas/MyMavenTask.java
@@ -1,4 +1,0 @@
-package io.github.fvarrui.javapackager.ideas;
-
-public class MyMavenTask implements MySettings{
-}
